refactor: unify ACL-rule parsing and errors into core/lib/general/acl-rules
setup/src/main.ts and run/src/main.ts each maintained their own copy of
buildACLRules/parseRulesOrThrow and the ACLRules/BuildACLRulesInput
interfaces. Move them to a shared module, and replace the INVALID_RULES
code duct-taped onto SetupError/SandboxError with a dedicated
InvalidRulesError (extends ActionError), letting callers narrow with
instanceof instead of relying on a shared string literal across two
otherwise-independent error-code unions.

Also share VerifyImageIdentity/VerifyImageDigestOptions/ResolvedImage
from core/lib/provenance/verify-image.ts instead of each entry point
redeclaring its own near-identical resolveVerifiedImage option/return
types.
