+++ setup/src/lib/errors.ts
@@ -3,13 +3,13 @@ import { ActionError } from "../../../core/lib/general/action-error.ts";
 /**
  * SetupError — intentional error in the setup action's own logic. Image
  * provenance failures throw ProvenanceError instead (see
- * core/lib/provenance/errors.js).
+ * core/lib/provenance/errors.js); invalid ACL rule syntax throws
+ * InvalidRulesError instead (see core/lib/general/acl-rules.js).
  *
  * Codes:
- *   INVALID_RULES        – ACL rule syntax error
  *   DOCKER_UNAVAILABLE    – docker CLI missing from PATH or a docker command failed
  *   INVALID_PROXY_ENGINE  – proxy_engine input isn't "transparent" or "explicit"
  */
-export type SetupErrorCode = "INVALID_RULES" | "DOCKER_UNAVAILABLE" | "INVALID_PROXY_ENGINE";
+export type SetupErrorCode = "DOCKER_UNAVAILABLE" | "INVALID_PROXY_ENGINE";
 
 export class SetupError extends ActionError<SetupErrorCode> {}
